Implementing request to hm-pricing

+++ src/main/java/com/home/manager/hmapplicationcore/dto/PropertyDto.java
@@ -16,5 +16,5 @@ public class PropertyDto {
     private String neighborhood;
 
     @NotEmpty
-    private BigDecimal size;
+    private Double size;
 }
